+++ src/views/SpaceDelegate.vue
@@ -297,7 +297,7 @@ onBeforeRouteLeave(async () => {
       <SpaceDelegatesDelegateRegistryV2Modal
         v-if="
           space.strategies?.some(
-            ({ name }) => name === DelegationTypes.DELEGATE_REGISTRY_V2
+            ({ name }) => name === DelegationTypes.SPLIT_DELEGATION
           )
         "
         :open="showDelegateModal"
